Fix: Position dialog after other citation tools
CiteFromIDDialog was sometimes appearing as the first
cite tool or as the last (before the basic and re-use
dialogs). Fixes so the dialog appears after the
citation tools but before the basic and re-use dialogs.

Also added ext.visualEditor.mwreference as a
dependency for ext.citoid.mwreference and fixed
case of ext.citoid.mwreference.

Bug: T87822
Change-Id: I7970cc656a639dfb1238e329baa9abbd82a695f0