Use 'rerender' event for generated content in citoid inspector
Replace the deprecated and evil 'DOMSubtreeModified' event and,
instead, use the generated node's 'rerender' event to update the
inspector that the widgets are rendered and ready.

We do this by creating a promise within the widget that resolves
when the widget is rendered. The inspector can then wait for the
promise to resolve before switching the ui to the results panel.

* Depends on ooui change I0248b47cca *
* Depends on ve-core change I2892409433b6d3 *

Change-Id: Iba36b4428c3ab8b704e8ecbafe66be2c188597a3