Allow structured citoid fields to be flattened
In the case where the field in the citation is
an Array [] or a 2D Array [[]], allow the values
to be concatenated if the templateData field is
of a lower dimension. At the inner level of the
2D arrays, values are concatenated with a space
as separator. At the higher level or in 1D arrays,
values are concatenated with a comma and a space.

Bug: T97044
Change-Id: I9746a36b951647bf40f66769290b064dd0b6d302