Disable input field on lookup, too
The lookup action is bound to the lookup input field, too (on key return),
so an user could submit the form twice (press return twice or click generate
button and press return), which would trigger an error message. To avoid that,
disable the input field, too, while the lookup processed.

Bug: T116471
Change-Id: I021c60094331d8816f20d3739f12726ce3df217a