Use basefields query parameter in citoid req
This change adds a basefields query parameter
to the request made to the citoid API (supported
in citoid versions 0.4.0 and higher- 0.4.0 is the
version currently in production).

In this version 0.4.0 of the citoid service, both the
base and the more specific types (i.e.
publicationTitle and websiteTitle) are returned
with this parameter. In later versions the
specific type will be removed and only the more
general type (i.e. publicationTitle) will be
returned.

Bug: T114515
Change-Id: Ia919fa4d1ef92989c5eed12c4d231fd1ca2bf684