ve.ui.CiteFromIdInspector: Add check for empty templates
In the buildTemplateResults function, a check has
been added for templates that match a citation
type but have no parameters for Citoid data. These
empty templates will now displa an error rather
than allowing the user to insert a blank citation.
The invalid templates are no longer added to the
final array of templates to be populated, and the
function rejects the promise if there were no
valid templates to be found.

Bug: T92428
Change-Id: I5cd5ca946ece43c050303d870bec4f35131bef36