Use new restbase endpoint for service
Switch to using new Restbase endpoint when it is
enabled in VisualEditor. In instances where Restbase
is not enabled in VisualEditor, it will continue
to use CitoidServiceUrl as supplied in extension.json

This also changes the request format from
format=mediawiki&basefields=true to the equivalent
of format=mediawiki&basefields=false, which is
just mediawiki in the new Restbase endpoint.

Bug:T114515
Bug:T152220
Change-Id: I78f2bc725e611ffea4932f134c66bca9c43d11e5