Let users know they can search by title or citation
Users have been able to search by title of the work
or using a full text citation via both worldcat (books)
and crossRef (things with dois including journal articles
and some academic books.) The feature also looks for a
url in the citation. This has been enabled since
August 2018.

This feature will not work if the work is not in
either of those databases or if the citation does
not contain a url.

Bug: T198567
Change-Id: I8f101b46c36bc4617ce3f9ba847c4dffa41133eb