Expose citoid map and conditionally register citefromid command
The citefromid command (which launches the inspector) is only made
available if some citation tools were exposed by the Cite extension
(this means that at least the 'manual' tab has some use, even if
'auto' is disabled).

Like the citation tool list, the citoid map is also exposed to
avoid code duplication.

As the citation tool list is truncated upstream (to 5) we don't
need to re-impose the limit here.

Bug: T219510
Change-Id: I90c4263b8eacf0039db9524d7ada4c598d508b6c
Depends-On: If848a7fbebf050cd1ba376e45c9be95a9ad98290