Merge 'ext.citoid.wikibase.data' into 'ext.citoid.wikibase.init'
Follows-up ebbeb70. The 'ext.citoid.wikibase.data' module is never used
or needed separately, and probably shouldn't be its own top-level
module bundle registered on all page views.

While at it, also adopt use of 'packageFiles' so that we don't need
use the client-side message transport for this at all, which is
otherwise not intended for content-language messages.

Change-Id: Ie787bb35a372839aba9229396949e819e4e2eea4