wb.Citoid.init: Remove fragile use of 'wikibase' global
When this module first executes, the 'wikibase' global might
not exist yet. This file has no dependencies (except for the
local data module that defines a JSON message). So there is no
reason to assume this would work.

Fortunately, it's not actually needed until after the using()
callback, which loads modules that define 'wikibase'.
Reference it there instead.

Change-Id: I7d3587211b50e58044a7f55eee8b484bbbc4f3d6