Don't add ext.citoid.wikibase.init to mobile pages
According to the extension.json definition this code should
not be loaded on the mobile site.

Instead of relying on the targets system to do this, it's better
to use the relevant service to do the check to make this more
explicit. The targets sytem will also be removed in future so
this future-proofs the code.

Bug: T324723
Change-Id: I7496f7ab049bfc424a6edeeb2ce475f13c38ce61