Add a panel to extend a reference (MMVP)
MMVP to allow users extending a reference by selecting a parent
from the of existing, extendable references.

I added a new tab to the CitoidInspector dialog. It's re-using the
SearchWidget to show the references that are part of internal list.

Selecting one will then open a new Dialog to set the content of
the extension. Ideally we would use a re-usable version of the
dialog that's also used to edit and create new references here.

The tab will not be visible when the feature flag for extending
references is disabled.

TODO
- Filter results in the extends search to leave out these that are
  already extensions

Bug: T365721
Bug: T367030
Depends-On: Idd579970cb64500dac27053213e9b116f23b6d76
Change-Id: Ife3b8dc78bf37377f42af53904947e746c3425a6