Reset search when closing the dialog
It does not make sense to keep whatever the user put in there.
Next time the dialog is openend they probably want to do something
else or try something else.

This is also already the case in the Cite extension and is also
done in the search field for the auto-generation.

Change-Id: I8f9df7d5c24dcbd2331238e8b98657b221bb5fd8