Use VE standard overlay inheritence for re-use sub-menu
Using the sub-menu in mobil mode was not possible due to the
positioning in relation to the rest of the interface. Using the VE
container overlay to make sure that the sub-menu is working in mobile
and desktop mode.

The default OOUI overlay selected by `$overlay: true` would be more
ideal and would work for an item at the bottom of the screen, but some
glitch perhaps with z-index causes bad behavior.

Bug: T375053
Change-Id: I62c5c0389183387d98c3c41b4e70df993b9ff6de