Use the ReferenceDialog for inserting a new sub-ref
The ReferenceDialog can be used to set the additional details
when creating a new sub-ref. It takes care of showing the edit
panel and actually inserting the ref.

Note also that the ReferenceDialog now not only takes care of
inserting the reference but also triggers the selection of it when
the dialog closes.

Bug: T375156
Depends-On: I47db21cd44d19ff209f9d5fa736bd6a057b983f2
Change-Id: Iae32e95b7ede1c9c0891a910537e1e36514f5a0c