Refactor error message layout to be a little clearer
Previously the code created a MessageWidget and reached into its
$element component to add some extra components – a second line of
error message and a call-to-action to fix things manually.

This change puts a custom element in the $label of the MessageWidget and
then exposes a new field, $customErrorLabel, to be set later. As a
result the contents of the MessageWidget are all vertically aligned
together, as a second visual column after the icon, rather than being
underneath the first line at the left of the element.

Bug: T379616
Change-Id: Ied1f9c1ec0534b90a00cd863008b8ad4397e0f82