Use restful api pattern for all configs
Native restful pattern from the citoid
api has been stable for a while now and
is now used in production behind api gateway.

(Restbase is now no longer used to serve citoid
requests in production.)

Switch to using restful pattern for CitoidServiceUrl;
users will either need to be running a version of the
citoid service of 1.2.0 and above, or be using
the $wgCitoidFullRestbaseURL and Restbase.

(Previously the query pattern was used instead.)

This will later allow seemless removal of the
$wgCitoidFullRestbaseURL config variable.

Also don't register automatic tab if no serviceUrl
is in the config.

Bug: T361576
Change-Id: I0941d6284c375ff0368d8e2eca76d5380e39ee03