Streamline fragile Cite tool unregistration
This doesn't fully solve the fragile dependency but improves it
quite a bit.

In detail:

1. Instead of iterating the names and waiting for specific names to
show up we can right away unregister by that name. There is no error
if a name doesn't exist. It just doesn't do anything then.

2. Luckily OO.Factory.unregister() works with both the name and the
actually registered class. Using the name is more straightforward
here since we are iterating the names anyway.

3. There is only one extra tool with the naming pattern "reference/".
It's much easier to find and ultimately more stable when we just
hard-code that instead of working with a vague pattern.

4. Instead of working with a class name (an abstract helper class
from Cite) we can work with the name, just like in all the other
cases in this loop. The naming pattern "cite-" is guaranteed and
already used in other places.

Bug: T400800
Change-Id: I95be2e41dd8aa0bfcebb7c59ee5de20df24c42e5