Fix "replace" behavior deleting unrelated nodes
Line #41 in this file restricts the "modelClasses" to instances of
ve.dm.MWReferenceNode. In case something else is selected (e.g. an
entire reference list) getSelectedNode returns null, as it should.

Still the code continues to replace the node (in this case the
reference list).

New behavior: The "replace" flag is only used when there is a valid
ve.dm.MWReferenceNode to replace.

We could do an extra instanceof check, but this already happens
inside of getSelectedNode, as explained above.

Note this is not a full fix for the issue described in T367910. This
minor bugfix just fixes a mistake in the code and prevents the most
obvious data loss. The UX (the behavior observable by the user) still
needs fixing.

This patch also inlines a pointless class property that's not used
anywhere else.

Bug: T367910
Change-Id: Id8ce62d9c9f820da591e77f47e392e62b483f38b