Fix inconsistent behavior when auto citations are not setup
One of the lines was updated in I0941d62 to depend on both config
flags, but the other line of code was not updated the same way. This
creates the situation where the dialog opens with the "auto" tab,
but I cannot use it. It always fails. When I switch to another tab
I cannot switch back to "auto" because it is disabled.

I'm also using the oportunity to do a little bit of closely related
code cleanup. The goal is to make the code more stable and robust to
not run into similar problems again.

Bug: T361576
Change-Id: I264cb1ca0dadc2706c89a9d9aa4e060666442b3a